refactor(openehr-adl): split adl14/convert.rs into converter core + three stage modules
convert.rs was the one file left over the ~1,000 non-test LOC bar. The three
stages that need no converter state move to siblings: walk (the read-only
definition traversals the code planning consumes, plus the shared
cco_data_mut accessor), multiplicity (the 1.4 default-occurrences
materialisation and RM-default elision) and metadata (the description /
standardised meta-data / HRID-version transform). Behaviour-preserving.
